requirements.txt
evaluate==0.4.2
openai==1.97.0
pandas==2.3.1
similarities==1.2.2
snowflake_connector_python==3.12.3
sqlparse==0.5.3
tqdm==4.66.2